Added gif, base presentation animator to library, readme.
